Introduces TagExtractor, for capturing binary annotations generically
This starts a decoupling between the mechanism to extract a tag, for
example the http path from a request, and actually wanting to add it to
a span. Once complete, it should be possible to add custom dimensions
like `servlet.user` without changing core code.

Current status: all http libraries now share a common value parser.
Next step: refactor out duplication and make it possible to configure
which tags to attach to the span.